requirements.txt
Flask==2.3.2
Requests==2.31.0

python-dotenv~=1.0.0